requirements.txt
streamlit==1.42.0
pandas==2.2.3
numpy==2.1.3
plotly==5.24.1
supabase==2.11.0
python-dotenv==1.0.1
pyarrow==19.0.0
fastparquet==2024.11.0

pyarrow
streamlit
